refactor(lesson-plan): move core AST to `Drasil/LessonPlan` namespace
