+++ code/drasil-docLang/lib/Drasil/ExtractLsnDesc.hs
@@ -7,7 +7,7 @@ import Drasil.Database (UID, ChunkDB)
 import Language.Drasil
 
 import Drasil.GetChunks (resolveBibliography)
-import Drasil.DocumentLanguage.Notebook.Core
+import Drasil.LessonPlan.Core
 import Drasil.ExtractCommon (extractChRefs)
 import Drasil.ExtractDocDesc (getSec)
 import Language.Drasil.Development (lnames)
